+++ skforecast/utils/tests/test_initialize_lags.py
@@ -74,9 +74,10 @@ def test_TypeError_initialize_lags_when_lags_is_not_valid_type_ForecasterAutoreg
     Test TypeError is raised when lags is not a valid type in ForecasterAutoregMultiVariate.
     """
     lags = 'not_valid_type'
+    
     err_msg = re.escape(
-        f"`lags` argument must be a dict, int, 1d numpy ndarray, range, tuple or list. "
-        f"Got {type(lags)}."
+        (f"`lags` argument must be a dict, int, 1d numpy ndarray, range, tuple or list. "
+         f"Got {type(lags)}.")
     )
     with pytest.raises(TypeError, match = err_msg):
         initialize_lags(